app/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
sqlalchemy==2.0.14
psycopg2==2.9.5
pydantic==1.10.6
jose==3.3.0
passlib==1.7.4
python-dotenv==0.21.0
pytest==7.2.2
pytest-asyncio==0.20.3
alembic==1.9.3
asyncpg==0.27.0
databases==0.7.4
